[FIX] spreadsheet_sale_management: field sync in multi model
Fix an issue where field sync would crash when used with UNDO/REDO.

The error came from the owl behviour of action created component:
The newly created component is created and started before unmounting
the old one. This disallows us to add and remove items from a registry
based on the component lifetime as they would be added to the registry by
the new component before being removed by the old one.

Opening the version history in another tab allow to bypass this restriction
as each tab has it's own registry of plugins/commands/... .

We also tried fixing the issue by adding Ids to the registry key but this
cannot be done as they would define the same getter twice.

closes odoo/enterprise#97424

Task: 4854879
X-original-commit: a9fd6648eaee5818af004a745f13468d9b708d63
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>