[FIX] account_accountant: normalised label
This commit will change the way we normalise the label. We now firstly check
that the label is a structure reference and if so we don't escape the numbers.
Otherwise, we do an escape of the label and the numbers.
Escaping the label will allow to avoid case where we use character like * or .
that would mess up the match regex.

Example:
'+++344/0660/16938+++' will be \\+\\+\\+344\\/0660\\/16938\\+\\+\\+
'This is a test' will be 'This\\ is\\ a\\ test'

task-5023114

closes odoo/enterprise#92973

X-original-commit: f3e170ce3af02c35e54c48c72128c3b904e5b4f8
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>