[IMP] account_reports, l10n_*: Return states refactoring
Before this commit, adding specific states to a return was tedious as it was required to add a field in Python and add overrides in the view using t-set, ...

Now everything is done on the Python side we do not have to change the view anymore when adding new states.
This is dynamically computed based on the result of '_get_state_field' and the current_state.

Also we noted an important thing, we should never set the state from the base field directly but by using the 'state' field. This will automatically trigger the write on the corresponding field and ensure everything is synchronized. This is because we do not want to have to add every new state field in the depends of the compute as it is error prone; we might forget it.
Using a common way of writing the state is easier.

task-5012501

closes odoo/enterprise#92618

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>