[IMP] project_enterprise: remove highlight on planned tasks
Before this commit, when the user uses the smart schedule on tasks to
schedule his tasks, a filter is added to highlight the tasks planned by
the smart schedule. The problem is the user would prefer directly having
an overview of all tasks planned instead of only see the ones planned.

This commit removes the highlight tasks planned by the smart tasks to
let the user have an overview of his tasks planned.

closes odoo/enterprise#92654

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>