[FIX] l10n_co_dian: prevent error when sending invoice in CO localization
Currently, an error occurs when the operation type (CO) is
"Nota Débito que referencia una factura electrónica" and, after confirmation,
the user attempts to send the mail.

Steps to Reproduce:
- Install Accounting and l10n_co_dian modules.
- Switch company to "CO Company".
- Create new customer invoice.
 (e.g;  Operation Type (CO) = Nota Débito que referencia una factura electrónica)
- Click on "Send" button. Ensure DIAN is selected in template and then send it.
- Error occurs.

Error:
AttributeError - 'bool' object has no attribute 'isoformat'

Cause:
The issue happens because debit_origin_id is not set, which makes
reference_invoice equal to None, leading to an error. - [1]

In pervious versions, the operation type could not be modified because it
was read-only field. From saas-18.4, it became a stored field. Therefore,
without debit note, the operation type cannot be changed directly.
In this case, the condition at [2] fails, and the method returns None.

Fix:
This commit ensures that a value is only returned if a reference invoice exists;
otherwise, it returns None.

[1] - https://github.com/odoo/enterprise/blob/8c0217e4e38903783ab7d70ca3eec1c1e3de03de/l10n_co_dian/models/account_edi_xml_ubl_dian.py#L730
[2] - https://github.com/odoo/enterprise/blob/b67a58ffd70952a06b7dd56a54782c433ff673fb/l10n_co_dian/models/account_edi_xml_ubl_dian.py#L1495-L1496

sentry-6810908763

closes odoo/enterprise#95400

X-original-commit: 4fcf436fcd4c8181b4671871f82e3dac1e753262
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>
Signed-off-by: Aditi Patel (adip) <adip@odoo.com>