[FIX] l10n_de_reports: fix datev account for sale, refund entries
In _get_datev_account, the l10n_de_datev_main_account_id is determined by identifying a singular debit or credit account used in the journal entry. If there’s no unique account, it falls back to searching for a unique non-tax line among debit or credit lines.

In POS, however, there is the possibility of generating journal entries that break this flow:

- In a single POS session, add product A with tax 19% and product B with tax 7%.

- In the same session, refund product A.

- Close the session to generate the entries.

In the resulting entry, since both sale and refund are present, it is not possible to discriminate using debit and credit amounts alone, and as a result, the field l10n_de_datev_main_account_id is not populated.

Since account 1411 is always the one to be used for l10n_de_datev_main_account_id in this specific case, this commit adds a final fallback filter to select the correct line and ensure the field is populated.

opw-4915272

closes odoo/enterprise#94421

X-original-commit: 13832d80570956e504e1c09f41acbeb0bc4baedc
Signed-off-by: Paolo Gatti (pgi) <pgi@odoo.com>
Signed-off-by: Guillaume Teboul-Tornezy (gute) <gute@odoo.com>