[IMP] shopfloor: show assigned employee on card
In shopfloor, to know which WOs are free to take,
Assigned employees are now shown on the card
(limited to 13 employees).

With the recent shopfloor redesign, it became less easy to identify active WOs.
Therefore, now any active WOs are highlighted for any user.

In overview, for the employee currently selected,
the WO quick access button is highlighted in WOs they're assigned to.

closes odoo/enterprise#92673

Task: 4987845
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>