[FIX] l10n_br_avatax: only include itemCode when needed
The PR #90703 introduced a mocking system for tests in the br modules,
however, there was an issue with this in the expected requests
validation. l10n_br_edi extends `_prepare_l10n_br_avatax_document_line_service_call`
to always include the `itemCode` key in the Avatax or Edi request,
as such this causes test cases to fail depending on module installation
state. If _edi is installed, it expects the request to contain itemCode,
if it isn't installed, the request is expected to not contain this key.
A fix for this is to check the state of the current database and see if
the module is installed, and include or remove that key from the json as
necessary.

This fix appears to be the smallest in terms of diff to solve the issue,
one of the other options is to override test cases in l10n_br_edi to
support the new key but that means that any new test cases that uses
this in the future will need to duplicate their code to _edi as well.

We may readdress this in the future with different approach but this
solves the nightly runbot issues from single app test.

runbot-230967

closes odoo/enterprise#92695

Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>