[IMP] hr_payroll: Introduce payslip correction mechanism
This commit introduces a correction mechanism for payslips to handle scenarios where a new employee version applies to a period for which a payslip was generated using an older version.

The existing "refund" action is renamed to "revert" and is now part of this broader correction process. A new payslip, reflecting the correct employee version, is automatically generated after the reversion.

Additionally, UX is improved with smart buttons on payslips, allowing users to easily navigate between the original, reverted, and corrected payslip versions.

closes odoo/enterprise#88961

Task: 4731029
Related: odoo/upgrade#8103
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>