[FIX] l10n_mx_hr_payroll: prevent error when creating payslips
Currently, an error occurs when creating a payslip for an employee.

Steps to Reproduce:
 - Install the l10n_mx_hr_payroll module with demo data.
 - Switch to the Mexico company.
 - Go to Employees and create an employee.
 - Click the Payslips button under the employee form.

TypeError: '>' not supported between instances of 'bool' and 'datetime.date'

This error occurs after [this commit](https://github.com/odoo/enterprise/commit/780a242b20932f808639f2719a9337d63ca6b13e#diff-e7b04636c3505b4f658eef73da77ef7dbd064ce663fdcfb76ed81e47246acc9d),
which replaces the use of the date_start and date_end fields with
contract_date_start and contract_date_end, since date_start and date_end
are deprecated. When accessing the version, the _compute_dates method
runs and sets date_start[1] to the version’s date if the version does not have
a contract_start_date. Before this commit, it always retrieved the version’s
date_start, which was handled safely. However, after this commit, it
retrieves contract_start_date instead of date_start, which is not a
required field and can be empty. As a result, when a user tries to create a
payslip, it raises the error [2].

This commit ensures that if the version does not have a contract_start_date,
then while computing the payslip amount it uses the version’s date_version,
which is a required field, thereby maintaining the same flow as in
previous versions.

[1]: https://github.com/odoo/odoo/blob/2054a54a66d6291ab37fa79f749a805138ada1c0/addons/hr/models/hr_version.py#L517-L521

[2]: https://github.com/odoo/enterprise/blob/f1c2326137fd3a94d37e2bbf09b175a9c742f46b/l10n_mx_hr_payroll/models/hr_payslip_worked_days.py#L21

sentry-6818787136

closes odoo/enterprise#93027

X-original-commit: 0dc4411f2756d5559348a9a8c1191c28c9aa6848
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Ashutosh Sharma (assh) <assh@odoo.com>