[IMP] documents: avoid warning during test for fake PDF
We modify Documents._get_is_multipage to emit an info instead of a warning
when in test mode and encountering a documents with pdf mime type but
containing instead html. We do that because
IrActionsReport._pre_render_qweb_pdf render pdf as html in test mode leading to
a lot of errors in the tests (because having documents with mimetype pdf but
containing html data).

Task-4873102

Part-of: odoo/enterprise#90662
Related: odoo/odoo#219925
Related: odoo/upgrade#8105
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>