[FIX] web_studio: reportEditor: do not display Label when inserting a X2M table
Click on the powerbox item that opens the popover to insert a x2many table.

Before this commit, the popover displayed the "Label" of the field (that is, in the case of a scalar or m2o field)
In a table this label is irrelevant.

After this commit, we hide the "Label" of the popover, the column takes the name of the field.

task-5017334

closes odoo/enterprise#92731

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>