[IMP] mail,*: do not use return value of store.insert
*: cloud_storage,im_livechat

Before this commit, results of the insertion of data in the store were
used, this could lead to issues later on, since the insertion does not
keep the order of the data.

To avoid this, we send back in the result of the rpc's the IDs linked to
the request.

In the future, we could use this to batch requests together and retrieve
the result of the request with the given ID.

task-4982525

closes odoo/enterprise#92738

Related: odoo/odoo#223412
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>