[IMP] documents_account: fix test account:TestChartTemplate.test_change_coa
Following the activation by default of the document synchronization, the test
account:TestChartTemplate.test_change_coa fails because a document is created
from an attachment already linked to an archived document.

How to reproduce: run the test account:TestChartTemplate.test_change_coa with
documents_account installed.

You get the following error:
duplicate key value violates unique constraint
"documents_document_attachment_unique"

This is because the documents exists but is archived so
_update_or_create_document creates a new document with the same attachment
leading to 2 documents pointing to the same attachment (this happens when
loading data from chart_template._load_data).

The archived document is created when AccountChartTemplate._load calls
records.with_context({MODULE_UNINSTALL_FLAG: True}).unlink() on account_move
which doesn't delete the related document because DocumentsUnlinkMixin archive
it instead to prevent the deletion of the attachment / documents when deleting
a record linked to the attachment (through res_model, res_id of the
attachment). So this cleaning phase is not complete.

We haven't found a real use-case where an attachment of an archived document is
used to create an account move but there is a similar use-case happening which
involve other part of the code:
- (Configure document synchronization for the sale journal)
- Create an invoice and send it
- Add an attachment
- Archive the invoice document
- Return to the invoice and set the second attachment as the main (it will
create a new document)
- Then change again to the previous one (you will get an error because it
tries to use the attachment of the archived document for another document)

So despite having no real use-case, we have decided in that case to update and
unarchive the document. We add a test that use an action to create an account
move from an archived document, even if that action is not present when
selecting a document from the trash.

Note: while refactoring the test a bit, we also correct the last assertion of
test_suspense_statement_line_id, checking that the partner is well transferred
to the account move.

Task-4873102

Part-of: odoo/enterprise#90662
Related: odoo/odoo#219925
Related: odoo/upgrade#8105
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>