[ADD] ai_crm(_livechat): allow lead creation from AI livechats
Purpose:
--------
This commit introduces 2 new bridge modules, ai_crm and ai_crm_livechat.

The first module adds an AI topic that allows AI agents to create leads
when the customer asks for more specific information about products,
services, or wants to get contacted back.
This topic comes with 2 tools:
- the first allows to get the available lead parameters that the agent can
chose from (currently teams and tags)
- the second allows to create the lead, after having asked the customer's
name, email and phone number.
To link leads with the agents that created them, the agent model will
inherit from utm.source.mixin.

The second module adds this topic to the AI livechat agent (this agent is
now a data record instead of a demo record).
Leads can now be created by this agent when a customer chats with it.

When chatting with an agent, some info about the current user will now be
sent to the llm (name, email, phone and job position). It has been added
so that the llm does not ask for them when creating a lead for a logged in
partner. It has been added in the `ai` module as it may be useful in other
scenarios too.

Task-5022760

Part-of: odoo/enterprise#92748
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>