[FIX] project_mrp_workorder_account: remove unecessary group from test
on the test `test_mrp_aa_employee_without_account_rights` the user was created
with the group `hr_timesheet.group_hr_timesheet_approver` which is not needed
and was causing the test to fail on Single app test as this module does not
depend on hr_timesheet.
This commit removes this group from the user creation.

runbot-231137

closes odoo/enterprise#92897

X-original-commit: 85d0c8b0c85d4ed3586c2b7b1ad20b7df554d385
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>
Signed-off-by: Abdelouahab Laaroussi (abla) <abla@odoo.com>