[IMP] mail,*: composer text to HTML draftContent
This commit updates the text field in composer model to use HTML
instead of plain text. This change is made to ensure that the html
composer can share the same field as the text composer.

Also, this commit renames the text field to draftContent
in a variety of places.

closes odoo/enterprise#92669

Related: odoo/odoo#223397
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>