[IMP] accountant_knowledge: improve formatting of template variables
This commit enhances the formatting of template variables in audit reports:
(1) Dates are now displayed according to the user's language settings (2)
Amounts now include the currency and are formatted based on the user's
preferences. By default, the company currency associated with the audit
report will be used.

Task-4989809

Part-of: odoo/enterprise#92765
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>