[FIX] accountant_knowledge: use correct company for template variables
When generating template variables for audit reports, the system currently
uses the company defined in the environment (i.e., `request.env`) instead
of the company linked to the audit report. This can lead to incorrect
amounts being displayed in the template variables. To resolve this, we
will explicitly compute the amounts using the company associated with
the audit report.

Task-4989809

Part-of: odoo/enterprise#92765
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>