[IMP] account_reports: translated return names
With this commit, return names are now translated to all languages
available at creation.

Another improvement is that the checks are also translated to all
available languages, this includes the name, the message and the
record name if there is. This is also supported using the check templates.

task-4818751

closes odoo/enterprise#92955

Related: odoo/upgrade#8328
Signed-off-by: Logan Staelens (lost) <lost@odoo.com>