[IMP] l10n_cl_edi{,_stock}: support delivery cross reference documents
Currently Delivery Guides can reference a Purchase Order by using the
client_order_ref field on the sale order linked to the picking. This
works but has issues, it is hardcoded to only that one reference
document, it isn't editable if needed, and has a hardcoded date of the
picking.

This PR takes advantage of the `l10n_cl.account.invoice.reference` model
that exists in `l10n_cl_edi` for `account.move` and supports identical
functionality on `stock.picking`.

It will automatically copy the reference from the sale order to the
picking if it is set and copy over to any related backorders. To remove
duplication, l10n_cl_edi was modified to extract the report table to a
reusable template that is accessible by both modules.

task-4970185

closes odoo/enterprise#92242

Related: odoo/upgrade#8222
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>