[IMP] timesheet_grid: don't cache list and kanban of timesheet with timer
Before this commit, since now the list view rpcs are cached in
timesheets, when the user stops the timer in another view and go back to
list view, the timer is still showed as running because of the values in
cache.

This commit disables the cache feature on list view of timesheet when
using the timer to avoid having desynchronisation.

task-4284216

Part-of: odoo/enterprise#81686
Related: odoo/odoo#202313
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>