[FIX] l10n_mx_hr_payroll: fix ISR when payslip overlaps 2 months
Bug:
When a payslip is overlapping two months, the ISR can be negative and is not
correctly computed.

Cause:
There is an adjustment made at the end of each month and each year. This works
only if the payslip ends on the last day of the month.

Fix:
We remove the adjustment and add an input for adjusting the ISR manually.

Task: 4790249
Part-of: odoo/enterprise#88167
Related: odoo/odoo#222754
Related: odoo/upgrade#8365
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>