[IMP] account_reports: Overrides 'return_periodicity' options when needed.
Before when we would have two return type pointing to the same report, the report options would not generate the 'return_periodicity'.

Now we added a way to pass manually the 'return_periodicity' option from the previous options. This is useful for instance when we open the report from the return as we know which return we target.

closes odoo/enterprise#92823

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>