[FIX] pos_appointment: phone number field empty
Task: [#5005216](https://www.odoo.com/odoo/project/1737/tasks/5005216)

---

The phone number field was not being populated correctly when booking a table from the website.
It is now properly filled using the phone number of the `res.partner` associated with the appointment.
The field remains unlinked to the partner, as it is just a copy.

closes odoo/enterprise#98041

X-original-commit: 77881bb9273913e2c11ed257c87e1a3385230368
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: Ludo Traina (ltra) <ltra@odoo.com>