[IMP] mail: remove sidebar concept in thread actions
This was added for contextual elligibility of showing of action
in sidebar vs discuss top-right thread actions and chat window
actions.

This commit replaces the `sidebarActions` to explicit code in
`condition`. The slight difference of sequences with partition
has been considered with dynamic sequence based on conditions for
"sidebar".

This change also removes `partition` and `contextual action` that
were defined to work in tandem with `sidebar` concept.

closes odoo/enterprise#93210

Related: odoo/odoo#224252
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>