[IMP] l10n_gt_edi: prevent move fields computation on install
This commit adds `_auto_init` to the `account.move` object in the
Guatemalan EDI that prefills all computed stored fields in the move
object with null values. This prevents computation on those fields when
installing the module, and also prevents MemoryError on large databases.

In addition, the guatemalan EDI will not be auto installed when the base
guatemalan localization is installed.

task-5031330

closes odoo/enterprise#93020

X-original-commit: f1ded135de3eaac43c89c0601fcc7faee0ba3647
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Yosua Nicolaus (yoni) <yoni@odoo.com>