[FIX] account_followup: only execute last followup once "per invoice"
Steps to reproduce:
- Install Accounting

- Go to "Accounting / Configuration / Invoicing / Follow-up Levels"
- Create a unique follow-up level:
  * Remind 1 day after due date
  * Send Email: [checked]
  * Automatic: [checked]
- Create an invoice:
  * Customer: [any]
  * Invoice Date: [1 month ago]
  * Due Date: [1 month ago]
  * Invoice Lines: [any]
- Confirm the invoice

- Go to "Settings / Technical / Automation / Scheduled Actions"
- Open "Account Report Followup; Execute followup"
- Run it manually
=> A reminder is correctly sent to customer (i.e. message on partner record)

- Wait for the following day
- Run the scheduled action again

Issue:
The same reminder in sent again.
The last reminder will be sent every day even if there is no new invoice.

Cause:
When computing the follow-up status for a contact, all the entries that are
due are taken into account.
The entries that have already been included in a follow-up should be ignored
when computing the same follow-up again.

Solution:
A field on a journal item (i.e. followup_line_id) tracks in which follow-up
level the item has been included.
We can exclude the journal items that has this field set to the next
follow-up level.

opw-4964103

closes odoo/enterprise#94045

X-original-commit: 8f3a09eaee93d8d74e38db9133fff6b51e62e18c
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>