[FIX] website_sale_renting_planning: correctly compute unavailability for rental product
Before this commit, when a service rental product linked to planning to
manage the hotel room booking for instance. When 2 shifts are "in
conflict" the calculation for the unavailability period for that product
is wrong. That compute will return the whole period of the both shifts
instead of returning the intersection of the both shifts (if only 2
resources are available for that product)

This commit reviews the calculation to make sure the quantity available
is correctly computed based on the shifts planned for that products.

task-4916650

Part-of: odoo/enterprise#93670
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>