[IMP] account_accountant: Change name of auto reco model
In this commit: https://github.com/odoo/enterprise/commit/e9ab2ae2876970c66c41fa56de70511af24471c1
we change the way the name of automatic reco model are made. But the solution we
choose are a bit naive. If the common substring is huge then the reco model is
unusable.
Since we are stuck with the constraint, the solution we choose is to add a
number in parentheses to avoid duplicate.

task-5031337

closes odoo/enterprise#93017

X-original-commit: 92b1d85719fd84a116bfcda2e93484b9d4f34462
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>