[REF] documents: fetch deletion delay when needed
`get_deletion_delay` is called every time the Documents app is open,
even though it's only needed when sending a document to the trash.

This adds network latency and uses resources for nothing.

With this commit, it's fetched only when needed: when the user sends a
document to the trash.
It's also cached in memory to avoid repeating the same RPC multiple times
if the users deletes multiple documents. The value is very unlikely to
change.

closes odoo/enterprise#92898

Signed-off-by: Florian Charlier (flch) <flch@odoo.com>