[FIX] sign: add document traceback
This commit fixes the problem introduced at odoo/enterprise#91434
where the add of documents in the Sign Template got broken by
a call to the undefined iframe of the document being added.

Additionally, another traceback was being thrown regarding
the saving of documents that were also depending on the
not yet rendered iframe. This was fixed by skipping the
call to the `saveChangesOnBackend` function in that not
yet ready iframes.

task-5032109

closes odoo/enterprise#92901

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>