[FIX] hr_payroll: remove join on version since it is usused in the query
Before this commit, the query used to get the bank account for each
employee, does a join on all version instead of just doing it on the
current version. Moreover, that join is not used at all inside the
query.

This commit completely removes the join on version in that query since
it seems useless to do that extra step.

Part-of: odoo/enterprise#92252
Related: odoo/odoo#222729
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>