[FIX] hr_payroll: avoid fetching all versions and filtering
This commit improves a bit the performance doing searching the versions
with a specific domain on contract_date_start instead of fetching all
versions and after that, filtering the versions based on
contract_date_start. Everything can now be done inside the search since
the contract date start is stored on db.

Part-of: odoo/enterprise#92252
Related: odoo/odoo#222729
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>