[FIX] hr*: payroll fields only accesssible by HR Admin
Before this commit, an HR Officer is able to access to payroll fields
like wage of each employee in the company. That behavior is unexpected,
only the HR manager (without hr_payroll module installed) should be able
to access to that information. The reason why the behavior before this
commit is unexpected is because before the HR refactoring there was an
addtional access right to access to contract. This access has been
removed and the access has been moved into HR manager group (when
`hr_payroll` module is not installed) or Payroll user when `hr_payroll`
module is installed.

Part-of: odoo/enterprise#92252
Related: odoo/odoo#222729
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>