[FIX] *hr_payroll: make sure groups on version fields are set on employee ones
This commit make sure groups applied on version fields are also applied
on inherited fields in `hr.employee`. By doing that, we could be sure
the fields are only accessible to hr manger (when `hr_payroll` module is
not installed) or payroll users (when `hr_payroll` module is installed).

Part-of: odoo/enterprise#92252
Related: odoo/odoo#222729
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>