[FIX] l10n_be_hr_payroll: add sudo to access to contract dates
Before this commit, when a HR user alters an employee, he will get an
access error because he does not access to contract_date_start field,
the reason is because we trigger the next activity of the current
version of an employee each time, the employee is altered. However, that
trigger is only done for the employee with an active contract (current
version with contract date start set). To do that, we need to do that
action in sudo since the contract dates are only accessible to payroll
user.

This commit adds a sudo to make sure the trigger next activity is done
without any issue after creating/altering an employee.

Part-of: odoo/enterprise#92252
Related: odoo/odoo#222729
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>