[FIX] hr_payroll: sudo to access some version fields
Before this commit, when a hr user alters/creates an employee, he could
get an access error because we need to have access to some fields in
version to recompute payslips for instance.

This commit makes sure the recomputation of the payslips can be done
after any changes making that recomputation needed without checking if
the user has access to certain fields in version.

Part-of: odoo/enterprise#92252
Related: odoo/odoo#222729
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>