[IMP] sale_subscription: invoice full period for  post paid products
Currently when users are manually invoicing a post-paid subscription in advance (before the next invoice date), the invoice is generated for the past period until today --> Only a partial period. This has been done to avoid users to invoice a future period during which some other products may be delivered.

taskid-4929802

Part-of: odoo/enterprise#93230
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>