[FIX] sale_subscription: Copy invoice journal on upsell or renewal
Up until now, the invoice journal set on a given subscription sale order
was not copied when the order was upsold or renewed. Only the
sale_order_template_id was copied, which depends on the company and has
a value for the invoice journal inside. Usually, the copany either
doesn't have a sale_order_template attached or they have one for normal
sale orders, not subscriptions. In these cases, the invoice journal was
extracted from the sale_order_template which didn't contain any (or any
valid) and so was left empty.

Here we check if the invoice journal value on the order has a value of
its own and, in that case, we copy it so that it is preserved. This
might be overridden by the value in the company template if there is
one.

closes odoo/enterprise#92981

Task: 5031427
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>