[FIX] iap_extract: fix unselection of boxes not saved to database
During refactoring of the boxes interface of the OCR (see commit
acfbaf3), this occurrence of `dataMoveId` wasn't replaced with
`recordId`.

It causes a small bug where the box unselection wouldn't work properly.
Visually, it looked like it was working as the JS code handling the
unselection of boxes worked fine, but the unselection wasn't saved to
the database. Upon refresh, all the boxes that were selected at some
point will still be displayed as "user selected" in the UI.

task-none

closes odoo/enterprise#93331

X-original-commit: db3a9cd394f0c7b2239a762b5f405ea15e8ebf83
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>