[FIX] mrp: correct total hours count in workcenter planning
Issue Before This Commit:
In the workcenter planning view, the total hours displayed were misleading,
as they always showed 24 hours of the day instead of the workcenter’s
defined working hours.
This caused confusion for users since the displayed total hours did not
match the actual available working time of the workcenter.

Steps to Reproduce:
- Create a Manufacturing Order (MO).
- Add "Drawer" as the product.
- Add a "SEC-ASSEM: [FURN_8855] Drawer" BOM.
- Set the quantity to produce as 10.
- Confirm the MO and plan it.
- Open the planning view and notice that the total hours count
  incorrectly shows 24 hours instead of the defined working hours.

Cause of the issue:
The bug was introduced by this PR (https://github.com/odoo/odoo/pull/205486),
as the related changes were not adapted here.
Because of that, pill.record.workcenter_id[0] returns an undefined value.

With This Commit:
The calculation of total hours has been corrected to consider only
the workcenter’s defined working hours.
Users now see accurate hours in the planning view, making it easier
to plan and schedule work orders reliably.

task - 4900885

closes odoo/enterprise#94859

X-original-commit: 936bfcf4aa0385b796a58cd30df6565f259e4328
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Divyanshu Hasmukhbhai Joshi (jodh) <jodh@odoo.com>