[FIX] helpdesk: order tickets by ticket_ref more intuitively
**Issue**
When using a `helpdesk.ticket` sequence with e.g. a padding of 2, once
reaching 100 tickets, ordering tickets by `ticket_ref` in the list view
is unintuitive as it is a Char field (so '11' > '100') and the results
are not useful if the user wants to see the oldest/newest tickets.

opw-4891916

closes odoo/enterprise#95971

X-original-commit: 1b39a7be8378d68bcf0b8461b3727350505be6c1
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>