[FIX] helpdesk: ensure partner belongs to the same company as the ticket
**Issue**:
It is possible to associate a partner that belongs to a different company than the ticket’s company. As a result, some users could see tickets in the list view that they should not have access to. When attempting to open those tickets, an error was raised. Even worse, in some cases (see the associated ticket), certain users were still able to access them.

**Steps to reproduce**:
- Create two companies (A and B)
- Switch to company A
- Open the Helpdesk application
- Create and open a ticket
- Select both companies A and B
- Assign a partner linked to company B to the ticket
- Save it

opw-4926497

closes odoo/enterprise#96350

X-original-commit: 34237e699813e590a2c0d1b4408270e5b086302e
Signed-off-by: Maxime Noirhomme (noma) <noma@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>