[IMP] planning,project: auto plan for flexible resources
Before this commit:
- auto plan, for planning and project, considers flex resources
as regular resources working with a regular calendar from 8 to 17.
for fully flexible resources, they are considered working the full
range, without checking leaves, and the overload on the day itself,
there is no dailt limit in the calendar, but the day is limited to 24
hours.

- when a task is assigned to many users, it's planned with its
real allocated hours, but the real time to allocated is allocated_hours
divided by the number of users.

After this commit:
- flex resources should not follow a regular calendar as flex
calendars doesn't have attendances, they can work at any moment
of the day without exceeding a number of hours per day and a
number of hours per week (configured in the flex calendar).
The auto plan plans shifts/tasks without exceeding daily/week rate
and without following a regular calendar, a flex employee is considered
working from 00:00 to 23:59 to cover more use-cases, such as restaurants,
hospitals, firefighters.
Same for fully flexible employees, the auto plan makes sure no overload
on a range of time, example: it's not possible to do a shift/task of 4hours
between 8 and 10 (only 2 working hours available)

- hours to plan for tasks = allocated_hours / number of users

task-4822880

closes odoo/enterprise#90948

Related: odoo/odoo#221253
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>