[PERF] iap_extract: remove savepoint when fetch OCR results
Very similar to commit 7f7ff8d, see full explanation in that commit
message.

This savepoint was also initiated in a loop which could cause
performances issues when too many of them are active at the same time.

This one is much less critical and can be removed without much trouble.
It was only allowing to recover from a serialization failure, in the
case where a user was concurrently editing a document while it was being
updated by the OCR.

closes odoo/enterprise#93114

X-original-commit: eaa0fbed68c3d7971b997d01d8c00e29ecd15b02
Signed-off-by: Thibault Francois <tfr@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>