[FIX] voip_ai: Ensure correct transcription status with error handling
Previously, if a call transcription request failed due to an oversized file, the `RequestEntityTooLarge`
exception would cause the entire transaction to roll back.
This meant that any attempt to set the `transcription_status` to an error state (e.g., "too_big_to_process")
would be reverted, leaving the call in a "pending" state and obscuring the actual reason for failure.

With this commit we don't raise error anymore.
Instead an appropriate status code is returned as an response, this way transcription status persists
yet client is notified.

related task
task-4532108

Part-of: odoo/enterprise#92954
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>